Fix Combined Rings breaking on Save
Fixed a bug where the mod was not working with the new Combined Rings from 1.5 and would set the player's Magnetism to the minimum if the player slept with one on, even if they took the ring off and put it on again.

Also, added compatibility with the new Glowstone ring.